Game refactoring

+++ Visualizer/Game.cpp
@@ -49,8 +49,8 @@ std::array<std::array<int, 8>, 8> Game::getCurrentDeskState() {
     mutex.lock();
 
     std::array<std::array<int, 8>, 8> res;
-    for (int i = 0; i < 8; ++i) {
-        for (int j = 0; j < 8; ++j) {
+    for (size_t i = 0; i < 8; ++i) {
+        for (size_t j = 0; j < 8; ++j) {
             res[i][j] = d(i, j);
         }
     }
